Add files via upload
This is the original upload of the main file. This will be known as version 1 and is subject to change. Every time you update the main file please update the date and the version on the top of the file before uploading the edit back onto github. If you make a new file follow the main files documentation style at the top and include a brief description of the file at the top of the file and in the commit changes (right here!).